requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
httpx
authlib
python-jose[cryptography]
itsdangerous  # Required for session middleware
pydantic[email]
pydantic-settings #v2
python-json-logger
slowapi
numpy
spacy